Fix mis-aligned movie tiles with associated 'More Info' buttons
The movie tiles were separated from their 'More Info' buttons. Created
a new div to wrap both the movie tile and the separate info button. It
was necessary to give the new containing div a class name of
'movie-tile' and give the actual movie tile a new class of
'movie-trailer-id' so that the modal function could connect properly.